+++ doc-examples/src/main/java/react4j/examples/dagger/MyComponent.java
@@ -1,8 +1,16 @@
 package react4j.examples.dagger;
 
+import javax.annotation.Nullable;
 import react4j.core.Component;
+import react4j.core.ReactNode;
 
 public class MyComponent
   extends Component
 {
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
 }